backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
openai>=1.12.0
python-dotenv>=1.0.1
alembic>=1.13.1
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
tenacity
